requirements.txt
openai
langchain
langchain_community
langchain_openai
python-dotenv
ipykernel
